+++ composeApp/src/commonMain/kotlin/ir/khebrati/audiosense/utils/CollectionOperators.kt
@@ -8,3 +8,9 @@ fun <T,R> Map<T,R>.copy(key: T, value: R): Map<T,R> {
         this[key] = value
     }
 }
+
+fun <K : Comparable<K>, V> Map<out K, V>.toSortedMap(): Map<K, V>{
+    val orderedKeys = this.keys.sorted()
+    val mappedValue = orderedKeys.map { this[it] }
+    return orderedKeys.zip(mappedValue).toMap() as Map<K, V>
+}
